src/FunctionApp/requirements.txt
azure-functions
pytest
python-dateutil
requests 